refactor(utils): split report_utils.py into a package
Convert python/pdstools/utils/report_utils.py (1317 lines) into a
package, mirroring the layout used for utils/cdh_utils. The public
surface is unchanged — every existing import (e.g.
"from pdstools.utils.report_utils import quarto_render_to_html")
continues to resolve, and dir(report_utils) is byte-identical.

New layout (all submodules underscore-prefixed; only __init__ is the
public surface):

* _common         — shared logger.
* _filenames      — output filename composition, report-resource copying.
* _quarto         — Quarto rendering, version detection, callouts,
                    show_credits.
* _html           — CSS inlining, zip bundling of Quarto resource folders,
                    report error scanning.
* _tables         — RAG-coloured metric tables (itables, great_tables).
* _polars_helpers — small polars helpers and report aggregations
                    (gains_table, hierarchy summarisations, etc.).
* _query          — pdstools QUERY serialise/deserialise.

__init__ re-exports both the public functions and a handful of
private helpers (_get_version_only, _set_command_options,
_write_params_files, _get_cmd_output, _inline_css) that the test
suite already pokes at.

A few tests in test_report_utils.py monkey-patched names on the
package module (report_utils.shutil, report_utils.quarto_print,
report_utils._get_cmd_output, report_utils.get_quarto_with_version,
report_utils.get_pandoc_with_version) and relied on those bindings
being the same objects the call sites used. After the split the
call sites live in report_utils._quarto with their own bindings, so
the patches are redirected to the actual module
(report_utils._quarto.<name>). No production code or test logic
otherwise changed.
