--- python/pdstools/utils/report_utils.py
@@ -1,1317 +1,0 @@
-import datetime
-import html
-import io
-import json
-import logging
-import os
-import re
-import shutil
-import subprocess
-import traceback
-import zipfile
-from pathlib import Path
-from typing import Any
-
-
-import polars as pl
-
-from ..utils.types import QUERY
-
-# Re-export RAG functions from metric_limits for convenience in Quarto reports.
-from .metric_limits import (  # noqa: F401
-    MetricFormats,
-    MetricLimits,
-    exclusive_0_1_range_rag,
-    positive_values,
-    standard_NBAD_channels_rag,
-    standard_NBAD_configurations_rag,
-    standard_NBAD_directions_rag,
-    standard_NBAD_predictions_rag,
-    strict_positive_values,
-)
-
-# Re-export NumberFormat, MetricFormats, and MetricLimits for external use
-from .number_format import NumberFormat  # noqa: F401
-
-logger = logging.getLogger(__name__)
-
-
-def get_output_filename(
-    name: str | None,
-    report_type: str,
-    model_id: str | None = None,
-    output_type: str = "html",
-) -> str:
-    """Generate the output filename based on the report parameters.
-
-    If ``name`` already ends with ``.{output_type}`` (case-insensitive), it
-    is treated as a full filename and returned verbatim (after replacing
-    spaces with underscores). Otherwise the filename is composed from
-    ``report_type``, ``name``, and ``model_id`` with the extension appended.
-    """
-    if name:
-        name = name.replace(" ", "_")
-        if name.lower().endswith(f".{output_type.lower()}"):
-            return name
-    if report_type == "ModelReport":
-        return f"{report_type}_{name}_{model_id}.{output_type}" if name else f"{report_type}_{model_id}.{output_type}"
-    return f"{report_type}_{name}.{output_type}" if name else f"{report_type}.{output_type}"
-
-
-def copy_quarto_file(qmd_file: str, temp_dir: Path) -> None:
-    """Copy the report quarto file to the temporary directory.
-
-    Parameters
-    ----------
-    qmd_file : str
-        Name of the Quarto markdown file to copy
-    temp_dir : Path
-        Destination directory to copy files to
-
-    Returns
-    -------
-    None
-
-    """
-    from pdstools import __reports__
-
-    shutil.copy(__reports__ / qmd_file, temp_dir)
-    shutil.copytree(__reports__ / "assets", temp_dir / "assets", dirs_exist_ok=True)
-
-
-def _write_params_files(
-    temp_dir: Path,
-    params: dict | None = None,
-    project: dict = {"type": "default"},
-    analysis: dict | None = None,
-    full_embed: bool = False,
-) -> None:
-    """Write parameters to YAML files for Quarto processing.
-
-    Parameters
-    ----------
-    temp_dir : Path
-        Directory where YAML files will be written
-    params : dict, optional
-        Parameters to write to params.yml, by default None
-    project : dict, optional
-        Project configuration to write to _quarto.yml, by default {"type": "default"}
-    analysis : dict, optional
-        Analysis configuration to write to _quarto.yml, by default None
-    full_embed : bool, default=False
-        When True, embeds all resources (JavaScript libraries like Plotly,
-        itables, etc.) for a fully standalone HTML (larger output).
-        When False, loads JavaScript libraries from CDN and skips esbuild
-        bundling (smaller output, but requires internet).
-
-    Returns
-    -------
-    None
-
-    """
-    import yaml  # type: ignore[import-untyped]
-
-    params = params or {}
-    analysis = analysis or {}
-
-    # Parameters to python code
-    with open(temp_dir / "params.yml", "w") as f:
-        yaml.dump(
-            params,
-            f,
-        )
-
-    # When not using full_embed, disable embed-resources so Quarto does not
-    # invoke esbuild to bundle JavaScript libraries (Plotly, itables, etc.).
-    # This avoids failures in environments where esbuild is unavailable
-    # (e.g. DJS Docker images that removed it due to CVE issues).
-    # See GitHub issue #620.
-    # plotly-connected: false = load Plotly from CDN (smaller file)
-    # plotly-connected: true = embed Plotly (larger file)
-    embed = full_embed
-    html_format: dict = {
-        "embed-resources": embed,
-        "plotly-connected": embed,
-    }
-
-    quarto_config: dict = {
-        "project": project,
-        "analysis": analysis,
-        "format": {
-            "html": html_format,
-        },
-    }
-
-    with open(temp_dir / "_quarto.yml", "w") as f:
-        yaml.dump(quarto_config, f)
-
-
-def run_quarto(
-    qmd_file: str | None = None,
-    output_filename: str | None = None,
-    output_type: str | None = "html",
-    params: dict | None = None,
-    project: dict = {"type": "default"},
-    analysis: dict | None = None,
-    temp_dir: Path = Path(),
-    *,
-    full_embed: bool = False,
-) -> int:
-    """Run the Quarto command to generate the report.
-
-    Parameters
-    ----------
-    qmd_file : str, optional
-        Path to the Quarto markdown file to render, by default None
-    output_filename : str, optional
-        Name of the output file, by default None
-    output_type : str, optional
-        Type of output format (html, pdf, etc.), by default "html"
-    params : dict, optional
-        Parameters to pass to Quarto execution, by default None
-    project : dict, optional
-        Project configuration settings, by default {"type": "default"}
-    analysis : dict, optional
-        Analysis configuration settings, by default None
-    temp_dir : Path, optional
-        Temporary directory for processing, by default Path(".")
-    full_embed : bool, default=False
-        When True, fully embeds all JavaScript libraries (Plotly, itables,
-        etc.) into the HTML output (larger file).
-        When False, loads JavaScript libraries from CDN and skips esbuild
-        bundling, avoiding the need for esbuild (see issue #620).
-
-    Returns
-    -------
-    int
-        Return code from the Quarto process (0 for success)
-
-    Raises
-    ------
-    RuntimeError
-        If the Quarto process fails (non-zero return code), includes captured output
-    subprocess.SubprocessError
-        If the Quarto command fails to execute
-    FileNotFoundError
-        If required files are not found
-
-    """
-
-    def get_command() -> list[str]:
-        quarto_exec, _ = get_quarto_with_version()
-        _command = [str(quarto_exec), "render"]
-
-        if qmd_file is not None:
-            _command.append(qmd_file)
-
-        options = _set_command_options(
-            output_type=output_type,
-            output_filename=output_filename,
-            execute_params=params is not None,
-        )
-
-        _command.extend(options)
-        return _command
-
-    if params is not None:
-        _write_params_files(
-            temp_dir,
-            params=params,
-            project=project,
-            analysis=analysis,
-            full_embed=full_embed,
-        )
-
-    # render file or render project with options
-    command = get_command()
-
-    logger.info("Executing: %s in temp directory %s", " ".join(command), temp_dir)
-
-    # Set QUARTO_PYTHON to ensure Quarto uses the same Python that's running pdstools.
-    # This is critical for isolated environments (uv tool, pipx) where the default
-    # system Python may not have the required dependencies like ipykernel.
-    import sys
-
-    env = os.environ.copy()
-    env["QUARTO_PYTHON"] = sys.executable
-    logger.info(f"Setting QUARTO_PYTHON to: {sys.executable}")
-
-    process = subprocess.Popen(
-        command,
-        stdout=subprocess.PIPE,
-        stderr=subprocess.STDOUT,  # Redirect stderr to stdout
-        cwd=temp_dir,
-        env=env,
-        text=True,
-        bufsize=1,  # Line buffered
-    )
-
-    # Capture all output for potential error reporting
-    output_lines: list[str] = []
-    if process.stdout is not None:
-        for line in iter(process.stdout.readline, ""):
-            line = line.strip()
-            output_lines.append(line)
-            logger.info(line)
-    else:  # pragma: no cover
-        logger.warning("subprocess.stdout is None, unable to read output")
-
-    return_code = process.wait()
-    message = f"Quarto process exited with return code {return_code}"
-    logger.info(message)
-
-    # Raise an exception with captured output if Quarto failed
-    if return_code != 0:
-        captured_output = "\n".join(output_lines)
-        # Surface common, actionable failure modes at the top of the message
-        # so they're not buried in hundreds of lines of Quarto output.
-        hints = []
-        permission_lines = [
-            ln for ln in output_lines if "permission denied" in ln.lower() or "access is denied" in ln.lower()
-        ]
-        if permission_lines:
-            hints.append(
-                "Permission denied while running Quarto. Common causes: the "
-                "output directory or temp directory is not writable, the Quarto "
-                "binary is locked by an antivirus scanner, or a previous run "
-                "left files open. Offending log lines:\n  - " + "\n  - ".join(permission_lines[:5])
-            )
-        hint_block = ("\nHint: " + "\n\nHint: ".join(hints) + "\n") if hints else ""
-        raise RuntimeError(
-            f"Quarto rendering failed with return code {return_code}.{hint_block}\nOutput:\n{captured_output}",
-        )
-
-    if not full_embed and output_type == "html" and output_filename is not None:
-        html_path = temp_dir / output_filename
-        if html_path.is_file():
-            n_inlined = _inline_css(html_path, temp_dir)
-            logger.info("Inlined %d CSS stylesheet(s) into %s", n_inlined, output_filename)
-
-    return return_code
-
-
-_LINK_STYLESHEET_RE = re.compile(
-    r"<link\b[^>]*\brel=[\"']stylesheet[\"'][^>]*>",
-    re.IGNORECASE,
-)
-_HREF_ATTR_RE = re.compile(r"\bhref=[\"']([^\"']+)[\"']", re.IGNORECASE)
-
-
-def _inline_css(html_path: Path, base_dir: Path) -> int:
-    """Inline relative CSS ``<link>`` tags in an HTML file.
-
-    Replaces each ``<link rel="stylesheet" href="...">`` whose ``href`` is a
-    relative path with an inline ``<style>`` block containing the CSS text.
-    Absolute URLs (``http://``, ``https://``, ``//``) are left untouched.
-    Missing files are logged as warnings and left alone.
-
-    Parameters
-    ----------
-    html_path : Path
-        HTML file to patch in-place.
-    base_dir : Path
-        Directory used to resolve relative ``href`` values.
-
-    Returns
-    -------
-    int
-        Number of CSS files successfully inlined.
-    """
-    content = html_path.read_text(encoding="utf-8")
-    inlined = 0
-
-    def _replace(match: re.Match) -> str:
-        nonlocal inlined
-        tag = match.group(0)
-        href_match = _HREF_ATTR_RE.search(tag)
-        if not href_match:
-            return tag
-        href = href_match.group(1)
-        if href.startswith(("http://", "https://", "//")):
-            return tag
-        css_path = (base_dir / href).resolve()
-        if not css_path.is_file():
-            logger.warning("CSS file not found, leaving <link> tag intact: %s", css_path)
-            return tag
-        css_content = css_path.read_text(encoding="utf-8")
-        inlined += 1
-        return f"<style>\n{css_content}\n</style>"
-
-    patched = _LINK_STYLESHEET_RE.sub(_replace, content)
-    if inlined:
-        html_path.write_text(patched, encoding="utf-8")
-        logger.debug("Inlined %d CSS file(s) into %s", inlined, html_path.name)
-    return inlined
-
-
-def _set_command_options(
-    output_type: str | None = None,
-    output_filename: str | None = None,
-    execute_params: bool = False,
-) -> list[str]:
-    """Set the options for the Quarto command.
-
-    Parameters
-    ----------
-    output_type : str, optional
-        Output format type (html, pdf, etc.), by default None
-    output_filename : str, optional
-        Name of the output file, by default None
-    execute_params : bool, optional
-        Whether to include parameter execution flag, by default False
-
-    Returns
-    -------
-    list[str]
-        list of command line options for Quarto
-
-    """
-    options = []
-    if output_type is not None:
-        options.append("--to")
-        options.append(output_type)
-    if output_filename is not None:
-        options.append("--output")
-        options.append(output_filename)
-    if execute_params:
-        options.append("--execute-params")
-        options.append("params.yml")
-    return options
-
-
-def copy_report_resources(resource_dict: list[tuple[str, str]]):
-    """Copy report resources from the reports directory to specified destinations.
-
-    Parameters
-    ----------
-    resource_dict : list[tuple[str, str]]
-        list of tuples containing (source_path, destination_path) pairs
-
-    Returns
-    -------
-    None
-
-    """
-    from pdstools import __reports__
-
-    for src, dest in resource_dict:
-        source_path = __reports__ / src
-        destination_path = dest
-
-        if destination_path == "":
-            destination_path = "./"
-
-        if os.path.isdir(source_path):
-            shutil.copytree(source_path, destination_path, dirs_exist_ok=True)
-        else:
-            shutil.copy(source_path, destination_path)
-
-
-def generate_zipped_report(output_filename: str, folder_to_zip: str):
-    """Generate a zipped archive of a directory.
-
-    This is a general-purpose utility function that can compress any directory
-    into a zip archive. While named for report generation, it works with any
-    directory structure.
-
-    Parameters
-    ----------
-    output_filename : str
-        Name of the output file (extension will be replaced with .zip)
-    folder_to_zip : str
-        Path to the directory to be compressed
-
-    Returns
-    -------
-    None
-
-    Raises
-    ------
-    FileNotFoundError
-        If the folder to zip does not exist or is not a directory
-
-    Examples
-    --------
-    >>> generate_zipped_report("my_archive.zip", "/path/to/directory")
-    >>> generate_zipped_report("report_2023", "/tmp/report_output")
-
-    """
-    if not os.path.isdir(folder_to_zip):
-        logger.error(f"The output path {folder_to_zip} is not a directory.")
-        return
-
-    if not os.path.exists(folder_to_zip):
-        logger.warning(
-            f"The {folder_to_zip} directory does not exist. Skipping zip creation.",
-        )
-        return
-
-    base_filename = os.path.splitext(output_filename)[0]
-    zippy = shutil.make_archive(base_filename, "zip", folder_to_zip)
-    logger.info(f"created zip file...{zippy}")
-
-
-def bundle_quarto_resources(output_path: Path) -> Path:
-    """Bundle a Quarto-rendered file with its resources folder into a zip.
-
-    When Quarto renders an HTML report without ``embed-resources``, it emits
-    the HTML alongside a ``<basename>_files/`` directory containing the
-    JavaScript and CSS assets the report needs. This helper detects that
-    pattern and wraps both into a single ``<basename>.zip`` archive so the
-    report can be distributed and unpacked as one unit.
-
-    If no companion resources folder exists next to ``output_path`` (e.g. the
-    report was fully embedded, or the format doesn't produce resources), the
-    function is a no-op and returns ``output_path`` unchanged.
-
-    Parameters
-    ----------
-    output_path : Path
-        Path to the rendered report file (typically an HTML file). The
-        companion resources folder is expected at ``<output_path stem>_files``
-        in the same directory.
-
-    Returns
-    -------
-    Path
-        Path to the zip archive when bundling occurred, otherwise the
-        original ``output_path``.
-    """
-    output_path = Path(output_path)
-    if not output_path.exists():
-        return output_path
-
-    resources_dir = output_path.with_name(f"{output_path.stem}_files")
-    if not resources_dir.is_dir():
-        return output_path
-
-    zip_path = output_path.with_suffix(".zip")
-    logger.info(
-        f"Bundling {output_path.name} with resources folder {resources_dir.name} into {zip_path.name}",
-    )
-    with zipfile.ZipFile(zip_path, "w", zipfile.ZIP_DEFLATED) as zf:
-        zf.write(output_path, output_path.name)
-        for file in resources_dir.rglob("*"):
-            if file.is_file():
-                zf.write(file, file.relative_to(output_path.parent))
-
-    shutil.rmtree(resources_dir, ignore_errors=True)
-    try:
-        output_path.unlink()
-    except OSError:  # pragma: no cover
-        pass
-    return zip_path
-
-
-def _get_cmd_output(args: list[str]) -> list[str]:
-    """Get command output in an OS-agnostic way."""
-    try:
-        result = subprocess.run(
-            args,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            text=True,
-            check=True,
-        )
-        return result.stdout.split("\n")
-    except subprocess.CalledProcessError as e:
-        logger.error(f"Failed to run command {' '.join(args)}: {e}")
-        raise FileNotFoundError(
-            f"Command failed. Make sure {args[0]} is installed and in the system PATH.",
-        )
-
-
-def _get_version_only(versionstr: str) -> str:
-    """Extract version number from version string."""
-    match = re.search(r"(\d+(?:\.\d+)*)", versionstr)
-    return match.group(1) if match else ""
-
-
-def get_quarto_with_version() -> tuple[Path, str]:
-    """Get Quarto executable path and version."""
-    try:
-        quarto_path = shutil.which("quarto")
-        if not quarto_path:
-            raise FileNotFoundError(
-                "Quarto executable not found. Please ensure Quarto is installed and in the system PATH.",
-            )
-        executable = Path(quarto_path)
-
-        version_string = _get_version_only(_get_cmd_output(["quarto", "--version"])[0])
-        logger.info("quarto version: %s", version_string)
-        return executable, version_string
-    except Exception as e:
-        logger.error(f"Error getting quarto version: {e}")
-        raise
-
-
-def get_pandoc_with_version() -> tuple[Path, str]:
-    """Get Pandoc executable path and version."""
-    try:
-        pandoc_path = shutil.which("pandoc")
-        if not pandoc_path:
-            raise FileNotFoundError(
-                "Pandoc executable not found. Please ensure Pandoc is installed and in the system PATH.",
-            )
-        executable = Path(pandoc_path)
-        if not executable:
-            raise FileNotFoundError(
-                "Pandoc executable not found. Please ensure Pandoc is installed and in the system PATH.",
-            )
-
-        version_string = _get_version_only(_get_cmd_output(["pandoc", "--version"])[0])
-        logger.info("pandoc version: %s", version_string)
-        return executable, version_string
-    except Exception as e:
-        logger.error(f"Error getting pandoc version: {e}")
-        raise
-
-
-def quarto_print(text):
-    from IPython.display import Markdown, display
-
-    display(Markdown(text))
-
-
-def quarto_callout_info(info):
-    quarto_print(
-        """
-::: {.callout-note}
-%s
-:::
-"""
-        % info,
-    )
-
-
-def quarto_callout_important(info):
-    quarto_print(
-        """
-::: {.callout-important}
-%s
-:::
-"""
-        % info,
-    )
-
-
-def quarto_plot_exception(plot_name: str, e: Exception):
-    quarto_print(
-        """
-::: {.callout-important collapse="true"}
-## Error rendering %s plot: %s
-
-%s
-:::
-"""
-        % (plot_name, e, traceback.format_exc()),
-    )
-
-
-def quarto_callout_no_prediction_data_warning(extra=""):
-    quarto_callout_important(f"Prediction Data is not available. {extra}")
-
-
-def quarto_callout_no_predictor_data_warning(extra=""):
-    quarto_callout_important(f"Predictor Data is not available. {extra}")
-
-
-def polars_col_exists(df, col):
-    return col in df.collect_schema().names() and df.collect_schema()[col] != pl.Null
-
-
-def polars_subset_to_existing_cols(all_columns, cols):
-    return [col for col in cols if col in all_columns]
-
-
-def create_metric_itable(
-    source_table: pl.DataFrame,
-    column_to_metric: dict | None = None,
-    column_descriptions: dict[str, str] | None = None,
-    color_background: bool = False,
-    strict_metric_validation: bool = True,
-    highlight_issues_only: bool = False,
-    rag_source: pl.DataFrame | None = None,
-    **itable_kwargs,
-):
-    """Create an interactive table with RAG coloring for metric columns.
-
-    Displays the table using itables with cells colored based on RAG
-    (Red/Amber/Green) status derived from metric thresholds.
-
-    Parameters
-    ----------
-    source_table : pl.DataFrame
-        DataFrame containing data columns to be colored.
-    column_to_metric : dict, optional
-        Mapping from column names (or tuples of column names) to one of:
-
-        - **str**: metric ID to look up in MetricLimits.csv
-        - **callable**: function(value) -> "RED"|"AMBER"|"YELLOW"|"GREEN"|None
-        - **tuple**: (metric_id, value_mapping) where value_mapping is a dict
-          that maps column values to metric values before evaluation.
-          Supports tuple keys for multiple values: {("Yes", "yes"): True}
-
-        If a column is not in this dict, its name is used as the metric ID.
-    column_descriptions : dict, optional
-        Mapping from column names to tooltip descriptions. When provided,
-        column headers will display the description as a tooltip on hover.
-        Example: {"Performance": "Model AUC performance metric"}
-    color_background : bool, default False
-        If True, colors the cell background. If False, colors the text (foreground).
-    strict_metric_validation : bool, default True
-        If True, raises an exception if a metric ID in column_to_metric
-        is not found in MetricLimits.csv. Set to False to skip validation.
-    highlight_issues_only : bool, default False
-        If True, only RED/AMBER/YELLOW values are styled (GREEN is not highlighted).
-        Set to False to also highlight GREEN values.
-    rag_source : pl.DataFrame, optional
-        If provided, RAG thresholds are evaluated against this DataFrame
-        instead of ``source_table``. Use this when ``source_table`` contains
-        non-numeric display values (e.g. HTML strings) but you still want
-        RAG coloring based on the original numeric data. Must have the same
-        columns and row order as ``source_table``.
-    **itable_kwargs
-        Additional keyword arguments passed to itables.show().
-        Common options include: lengthMenu, paging, searching, ordering.
-
-    Returns
-    -------
-    itables HTML display
-        An itables display object that will render in Jupyter/Quarto.
-
-    Examples
-    --------
-    >>> from pdstools.utils.report_utils import create_metric_itable
-    >>> create_metric_itable(
-    ...     df,
-    ...     column_to_metric={
-    ...         # Simple metric ID
-    ...         "Performance": "ModelPerformance",
-    ...         # Custom RAG function
-    ...         "Channel": standard_NBAD_channels_rag,
-    ...         # Value mapping: column values -> metric values
-    ...         "AGB": ("UsingAGB", {"Yes": True, "No": False}),
-    ...         # Multiple column values to same metric value
-    ...         "AGB": ("UsingAGB", {("Yes", "yes", "YES"): True, "No": False}),
-    ...     },
-    ...     column_descriptions={
-    ...         "Performance": "Model AUC performance metric",
-    ...         "Channel": "Communication channel for the action",
-    ...     },
-    ...     paging=False
-    ... )
-
-    """
-    from itables import show
-
-    RAG_COLORS = {
-        "RED": "orangered",
-        "AMBER": "orange",
-        "YELLOW": "yellow",
-    }
-    if not highlight_issues_only:
-        RAG_COLORS["GREEN"] = "green"
-
-    # Get RAG values using the shared function from metric_limits
-    from .metric_limits import add_rag_columns
-
-    df_with_rag = add_rag_columns(
-        rag_source if rag_source is not None else source_table,
-        column_to_metric=column_to_metric,
-        strict_metric_validation=strict_metric_validation,
-    )
-
-    # Convert to pandas for styling
-    pdf = source_table.to_pandas()
-    pdf_rag = df_with_rag.to_pandas()
-
-    # Rename columns to include HTML tooltips if column_descriptions provided
-    # Must be done before styling since Styler doesn't have rename method
-    column_rename = {}
-    if column_descriptions:
-        for col in source_table.columns:
-            if col in column_descriptions:
-                escaped_desc = html.escape(column_descriptions[col], quote=True)
-                column_rename[col] = f'<span title="{escaped_desc}">{col}</span>'
-        if column_rename:
-            pdf = pdf.rename(columns=column_rename)
-
-    def style_row(row):
-        styles = []
-        for col in pdf.columns:
-            rag_col = f"{col}_RAG"
-            if rag_col in pdf_rag.columns:
-                rag_val = pdf_rag.loc[row.name, rag_col]
-                if rag_val in RAG_COLORS:
-                    color = RAG_COLORS[rag_val]
-                    if color_background:
-                        styles.append(f"background-color: {color}")
-                    else:
-                        styles.append(f"color: {color}; font-weight: bold")
-                else:
-                    styles.append("")
-            else:
-                styles.append("")
-        return styles
-
-    # Build format dict from centralized config using MetricFormats
-    expanded_mapping = {}
-    for key, value in (column_to_metric or {}).items():
-        if isinstance(key, tuple):
-            for col in key:
-                expanded_mapping[col] = value
-        else:
-            expanded_mapping[key] = value
-
-    format_dict = {}
-    # When rag_source is provided, source_table may contain pre-formatted strings;
-    # use the numeric source for column detection and sort values, but only
-    # populate format_dict for columns that are still numeric in source_table.
-    numeric_source = rag_source if rag_source is not None else source_table
-    numeric_cols = []
-    for col in source_table.columns:
-        if not numeric_source[col].dtype.is_numeric():
-            continue
-        numeric_cols.append(col)
-        metric_id = expanded_mapping.get(col, col)
-        if isinstance(metric_id, tuple):
-            metric_id = metric_id[0]
-        if isinstance(metric_id, str):
-            metric_fmt = MetricFormats.get(metric_id) or MetricFormats.DEFAULT_FORMAT
-            if source_table[col].dtype.is_numeric():
-                format_dict[col] = metric_fmt.to_pandas_format()
-        else:
-            if source_table[col].dtype.is_numeric():
-                format_dict[col] = MetricFormats.DEFAULT_FORMAT.to_pandas_format()
-
-    # Add hidden sort columns for all formatted numeric columns so that
-    # DataTables can sort by the raw numeric value instead of the display string
-    # (e.g. "4K" < "500" or "100%" < "2%" lexicographically without this).
-    sort_col_map = {col: f"__sort__{col}" for col in numeric_cols}
-    for col, sort_col in sort_col_map.items():
-        pdf[sort_col] = numeric_source[col].to_pandas()
-
-    styled_df = pdf.style.apply(style_row, axis=1).format(format_dict, na_rep="")
-
-    # Set default itable options
-    default_kwargs: dict[str, Any] = {
-        "paging": False,
-        "allow_html": True,  # Required for styled DataFrames
-        "classes": "compact",
-        # "maxBytes":0,
-        # "maxRows":0,
-        "connected": True,
-        "show_dtypes": False,
-    }
-    default_kwargs.update(itable_kwargs)
-
-    if sort_col_map:
-        final_cols = list(pdf.columns)
-        sort_defs = []
-        for orig_col, sort_col in sort_col_map.items():
-            display_name = column_rename.get(orig_col, orig_col)
-            display_idx = final_cols.index(display_name)
-            sort_idx = final_cols.index(sort_col)
-            sort_defs.extend(
-                [
-                    {"targets": display_idx, "orderData": [sort_idx]},
-                    {"targets": sort_idx, "visible": False, "searchable": False},
-                ]
-            )
-        default_kwargs["columnDefs"] = default_kwargs.get("columnDefs", []) + sort_defs
-
-    return show(styled_df, **default_kwargs)
-
-
-def create_metric_gttable(
-    source_table: pl.DataFrame,
-    title: str | None = None,
-    subtitle: str | None = None,
-    column_to_metric: dict | None = None,
-    column_descriptions: dict[str, str] | None = None,
-    color_background: bool = True,
-    strict_metric_validation: bool = True,
-    highlight_issues_only: bool = True,
-    **gt_kwargs,
-):
-    """Create a great_tables table with RAG coloring for metric columns.
-
-    Displays the table using great_tables with cells colored based on RAG
-    (Red/Amber/Green) status derived from metric thresholds.
-
-    Parameters
-    ----------
-    source_table : pl.DataFrame
-        DataFrame containing data columns to be colored.
-    title : str, optional
-        Table title.
-    subtitle : str, optional
-        Table subtitle.
-    column_to_metric : dict, optional
-        Mapping from column names (or tuples of column names) to one of:
-
-        - **str**: metric ID to look up in MetricLimits.csv
-        - **callable**: function(value) -> "RED"|"AMBER"|"YELLOW"|"GREEN"|None
-        - **tuple**: (metric_id, value_mapping) where value_mapping is a dict
-          that maps column values to metric values before evaluation.
-          Supports tuple keys for multiple values: {("Yes", "yes"): True}
-
-        If a column is not in this dict, its name is used as the metric ID.
-    column_descriptions : dict, optional
-        Mapping from column names to tooltip descriptions. When provided,
-        column headers will display the description as a tooltip on hover.
-        Example: {"Performance": "Model AUC performance metric"}
-    color_background : bool, default True
-        If True, colors the cell background. If False, colors the text.
-    strict_metric_validation : bool, default True
-        If True, raises an exception if a metric ID in column_to_metric
-        is not found in MetricLimits.csv. Set to False to skip validation.
-    highlight_issues_only : bool, default True
-        If True, only RED/AMBER/YELLOW values are styled (GREEN is not highlighted).
-        Set to False to also highlight GREEN values.
-    **gt_kwargs
-        Additional keyword arguments passed to great_tables.GT constructor.
-        Common options include: rowname_col, groupname_col.
-
-    Returns
-    -------
-    great_tables.GT
-        A great_tables instance with RAG coloring applied.
-
-    Examples
-    --------
-    >>> from pdstools.utils.report_utils import create_metric_gttable
-    >>> create_metric_gttable(
-    ...     df,
-    ...     title="Model Overview",
-    ...     column_to_metric={
-    ...         # Simple metric ID
-    ...         "Performance": "ModelPerformance",
-    ...         # Custom RAG function
-    ...         "Channel": standard_NBAD_channels_rag,
-    ...         # Value mapping: column values -> metric values
-    ...         "AGB": ("UsingAGB", {"Yes": True, "No": False}),
-    ...         # Multiple column values to same metric value
-    ...         "AGB": ("UsingAGB", {("Yes", "yes", "YES"): True, "No": False}),
-    ...     },
-    ...     column_descriptions={
-    ...         "Performance": "Model AUC performance metric",
-    ...         "Channel": "Communication channel for the action",
-    ...     },
-    ...     rowname_col="Name",
-    ... )
-
-    """
-    import html as html_module
-
-    from great_tables import GT, html, loc, style
-
-    from .metric_limits import add_rag_columns
-
-    RAG_COLORS = {"RED": "orangered", "AMBER": "orange", "YELLOW": "yellow"}
-    if not highlight_issues_only:
-        RAG_COLORS["GREEN"] = "green"
-
-    gt = GT(source_table, **gt_kwargs)
-    gt = gt.tab_options(
-        table_font_size="12px",
-        column_labels_font_size="12px",
-    )
-    gt = gt.sub_missing(missing_text="")
-
-    if title is not None:
-        gt = gt.tab_header(title=title, subtitle=subtitle)
-
-    # Apply column label tooltips if column_descriptions provided
-    if column_descriptions:
-        label_kwargs = {}
-        for col in source_table.columns:
-            if col in column_descriptions:
-                escaped_desc = html_module.escape(column_descriptions[col], quote=True)
-                # Wrap column label in span with title attribute for tooltip
-                label_kwargs[col] = html(f'<span title="{escaped_desc}">{col}</span>')
-        if label_kwargs:
-            gt = gt.cols_label(**label_kwargs)  # type: ignore[arg-type]  # great_tables overload routes **kwargs to first positional
-
-    # Expand tuple keys to individual columns
-    expanded_mapping = {}
-    for key, value in (column_to_metric or {}).items():
-        if isinstance(key, tuple):
-            for col in key:
-                expanded_mapping[col] = value
-        else:
-            expanded_mapping[key] = value
-
-    # Apply formatting based on metric type using MetricFormats
-    formatted_cols = set()
-    for col in source_table.columns:
-        metric_id = expanded_mapping.get(col, col)
-        if isinstance(metric_id, tuple):
-            metric_id = metric_id[0]
-        if isinstance(metric_id, str):
-            fmt = MetricFormats.get(metric_id)
-            if fmt is not None:
-                gt = fmt.apply_to_gt(gt, [col])
-                formatted_cols.add(col)
-        elif callable(metric_id):
-            # When a callable is used for RAG, still apply formatting if
-            # the column name itself has a format defined in MetricFormats
-            fmt = MetricFormats.get(col)
-            if fmt is not None:
-                gt = fmt.apply_to_gt(gt, [col])
-                formatted_cols.add(col)
-
-    # Apply default number formatting to numeric columns not yet formatted
-    # Exclude columns used as row/group names from numeric formatting
-    rowname_col = gt_kwargs.get("rowname_col")
-    groupname_col = gt_kwargs.get("groupname_col")
-    numeric_cols = [
-        col
-        for col in source_table.columns
-        if col not in formatted_cols
-        and col != rowname_col
-        and col != groupname_col
-        and source_table[col].dtype.is_numeric()
-    ]
-    if numeric_cols:
-        gt = MetricFormats.DEFAULT_FORMAT.apply_to_gt(gt, numeric_cols)
-
-    # Apply RAG coloring
-    df_with_rag = add_rag_columns(
-        source_table,
-        column_to_metric=column_to_metric,
-        strict_metric_validation=strict_metric_validation,
-    )
-
-    for col in source_table.columns:
-        rag_col = f"{col}_RAG"
-        if rag_col not in df_with_rag.columns or df_with_rag[rag_col].dtype == pl.Null:
-            continue
-
-        for rag_value, color in RAG_COLORS.items():
-            row_indices = df_with_rag.with_row_index().filter(pl.col(rag_col) == rag_value)["index"].to_list()
-            if row_indices:
-                if color_background:
-                    gt = gt.tab_style(
-                        style=style.fill(color=color),
-                        locations=loc.body(columns=col, rows=row_indices),
-                    )
-                else:
-                    gt = gt.tab_style(
-                        style=style.text(color=color, weight="bold"),
-                        locations=loc.body(columns=col, rows=row_indices),
-                    )
-
-    return gt
-
-
-def n_unique_values(dm, all_dm_cols, fld):
-    if not isinstance(fld, list):
-        fld = [fld]
-    fld = polars_subset_to_existing_cols(all_dm_cols, fld)
-    if len(fld) == 0:
-        return 0
-    return dm.model_data.select(pl.col(fld)).drop_nulls().collect().n_unique()
-
-
-def max_by_hierarchy(dm, all_dm_cols, fld, grouping):
-    if not isinstance(fld, list):
-        fld = [fld]
-    fld = polars_subset_to_existing_cols(all_dm_cols, fld)
-    if len(fld) == 0:
-        return 0
-    grouping = polars_subset_to_existing_cols(all_dm_cols, grouping)
-    if len(grouping) == 0:
-        return 0
-    return (
-        dm.model_data.group_by(grouping)
-        .agg(pl.col(fld).drop_nulls().n_unique())
-        .select(fld)
-        .drop_nulls()
-        .max()
-        .collect()
-        .item()
-    )
-
-
-def avg_by_hierarchy(dm, all_dm_cols, fld, grouping):
-    if not isinstance(fld, list):
-        fld = [fld]
-    fld = polars_subset_to_existing_cols(all_dm_cols, fld)
-    if len(fld) == 0:
-        return 0
-    grouping = polars_subset_to_existing_cols(all_dm_cols, grouping)
-    if len(grouping) == 0:
-        return 0
-    return (
-        dm.model_data.group_by(grouping)
-        .agg(pl.col(fld).drop_nulls().n_unique())
-        .select(fld)
-        .drop_nulls()
-        .mean()
-        .collect()
-        .item()
-    )
-
-
-def sample_values(dm, all_dm_cols, fld, n=6):
-    if not isinstance(fld, list):
-        fld = [fld]
-    fld = polars_subset_to_existing_cols(all_dm_cols, fld)
-    if len(fld) == 0:
-        return "-"
-    return (
-        dm.model_data.select(
-            pl.concat_str(fld, separator="/").alias("__SampleValues__"),
-        )
-        .drop_nulls()
-        .collect()
-        .to_series()
-        .unique()
-        .sort()
-        .to_list()[:n]
-    )
-
-
-def show_credits(quarto_source: str | None = None):
-    """Display a credits section with build metadata at the end of a report.
-
-    Prints a formatted block containing the generation timestamp, Quarto and
-    Pandoc versions, and optionally the source notebook path.
-
-    Parameters
-    ----------
-    quarto_source : str, optional
-        Path or identifier of the source .qmd file. Include this for
-        standalone reports where knowing the source is useful. Omit for
-        Quarto website projects where pages are generated from templates.
-    """
-    _, quarto_version = get_quarto_with_version()
-    _, pandoc_version = get_pandoc_with_version()
-
-    timestamp_str = datetime.datetime.now().strftime("%d %b %Y %H:%M:%S")
-
-    lines = [
-        f"Document created at: {timestamp_str}",
-    ]
-    if quarto_source:
-        lines.append(f"This notebook: {quarto_source}")
-    lines.extend(
-        [
-            f"Quarto runtime: {quarto_version}",
-            f"Pandoc: {pandoc_version}",
-        ]
-    )
-
-    quarto_print("\n\n    ".join([""] + lines) + "\n\n    ")
-
-
-def serialize_query(query: QUERY | None) -> dict | None:
-    if query is None:
-        return None
-
-    if isinstance(query, pl.Expr):
-        query = [query]
-
-    if isinstance(query, (list, tuple)):
-        serialized_exprs = {}
-        for i, expr in enumerate(query):
-            if not isinstance(expr, pl.Expr):
-                raise ValueError("All items in query list must be Expressions")
-            serialized_exprs[str(i)] = json.loads(expr.meta.serialize(format="json"))
-        return {"type": "expr_list", "expressions": serialized_exprs}
-
-    if isinstance(query, dict):
-        return {"type": "dict", "data": query}
-
-    raise ValueError(f"Unsupported query type: {type(query)}")
-
-
-def deserialize_query(serialized_query: dict | None) -> QUERY | None:
-    """Deserialize a query that was previously serialized with serialize_query.
-
-    Parameters
-    ----------
-    serialized_query : dict | None
-        A serialized query dictionary created by serialize_query
-
-    Returns
-    -------
-    QUERY | None
-        The deserialized query
-
-    """
-    if serialized_query is None:
-        return None
-
-    if serialized_query["type"] == "expr_list":
-        expr_list = []
-        for _, val in serialized_query["expressions"].items():
-            json_str = json.dumps(val)
-            str_io = io.StringIO(json_str)
-            expr_list.append(pl.Expr.deserialize(str_io, format="json"))
-        return expr_list
-
-    if serialized_query["type"] == "dict":
-        return serialized_query["data"]
-
-    raise ValueError(f"Unknown query type: {serialized_query['type']}")
-
-
-def gains_table(
-    df: pl.LazyFrame | pl.DataFrame,
-    value: str,
-    index: str | None = None,
-    by: str | list[str] | None = None,
-) -> pl.DataFrame:
-    """Calculate cumulative gains for visualization.
-
-    Computes cumulative distribution of a value metric, sorted by the ratio
-    of value to index (or by value alone if no index). Used for gains charts
-    to show model response skewness.
-
-    Parameters
-    ----------
-    df : pl.LazyFrame | pl.DataFrame
-        Input data containing the value and optional index columns
-    value : str
-        Column name containing the metric to compute gains for (e.g., "ResponseCount")
-    index : str, optional
-        Column name to normalize by (e.g., population size). If None, uses row count.
-    by : str | list[str], optional
-        Column(s) to group by for separate gain curves. If None, computes single curve.
-
-    Returns
-    -------
-    pl.DataFrame
-        DataFrame with columns:
-        - cum_x: Cumulative proportion of index (or models)
-        - cum_y: Cumulative proportion of value
-        - by columns: if `by` is specified
-
-    Examples
-    --------
-    >>> # Single gains curve for response count
-    >>> gains = gains_table(df, value="ResponseCount")
-
-    >>> # Gains curves by channel, normalized by population
-    >>> gains = gains_table(df, value="Positives", index="Population", by="Channel")
-    """
-    # Determine sorting expression
-    sort_expr = pl.col(value) if index is None else pl.col(value) / pl.col(index)
-
-    # Determine index expression for cumulative x-axis
-    if index is None:
-        index_expr = pl.int_range(1, pl.len() + 1) / pl.len()
-    else:
-        index_expr = pl.cum_sum(index) / pl.sum(index)
-
-    if by is None:
-        # Single gains curve
-        gains_df = pl.concat(
-            [
-                pl.DataFrame(data={"cum_x": [0.0], "cum_y": [0.0]}).lazy(),
-                df.lazy()
-                .sort(sort_expr, descending=True)
-                .select(
-                    index_expr.cast(pl.Float64).alias("cum_x"),
-                    (pl.cum_sum(value) / pl.sum(value)).cast(pl.Float64).alias("cum_y"),
-                ),
-            ]
-        )
-    else:
-        # Multiple gains curves grouped by column(s)
-        by_as_list = by if isinstance(by, list) else [by]
-        sort_expr_with_by = by_as_list + [sort_expr]
-        gains_df = (
-            df.lazy()
-            .sort(sort_expr_with_by, descending=True)
-            .select(
-                by_as_list
-                + [
-                    index_expr.over(by).cast(pl.Float64).alias("cum_x"),
-                    (pl.cum_sum(value) / pl.sum(value)).over(by).cast(pl.Float64).alias("cum_y"),
-                ]
-            )
-        )
-        # Add entry for the (0,0) point for each group
-        gains_df = pl.concat([gains_df.group_by(by).agg(cum_x=pl.lit(0.0), cum_y=pl.lit(0.0)), gains_df]).sort(
-            by_as_list + ["cum_x"]
-        )
-
-    return gains_df.collect()
-
-
-def check_report_for_errors(html_path: str | Path) -> list[str]:
-    """Check generated report HTML for error indicators.
-
-    Scans the HTML file for error patterns that indicate plot rendering failures
-    or exceptions during report generation. These errors are typically hidden in
-    collapsed callout sections but should be caught in testing.
-
-    Parameters
-    ----------
-    html_path : str or Path
-        Path to the HTML file to check
-
-    Returns
-    -------
-    list[str]
-        List of error descriptions found (empty if no errors)
-
-    Raises
-    ------
-    FileNotFoundError
-        If the HTML file does not exist
-
-    Examples
-    --------
-    >>> from pdstools.utils.report_utils import check_report_for_errors
-    >>> errors = check_report_for_errors("HealthCheck.html")
-    >>> if errors:
-    ...     print(f"Found {len(errors)} error(s):")
-    ...     for error in errors:
-    ...         print(f"  - {error}")
-    """
-    html_path = Path(html_path)
-
-    if not html_path.exists():
-        raise FileNotFoundError(f"HTML file not found: {html_path}")
-
-    try:
-        if html_path.suffix.lower() == ".zip":
-            with zipfile.ZipFile(html_path) as zf:
-                html_members = [n for n in zf.namelist() if n.endswith(".html")]
-                if not html_members:
-                    raise IOError(f"No HTML file found inside zip: {html_path}")
-                content = zf.read(html_members[0]).decode("utf-8")
-        else:
-            content = html_path.read_text(encoding="utf-8")
-    except Exception as e:
-        raise IOError(f"Failed to read HTML file: {e}")
-
-    errors = []
-
-    # Common error patterns in HTML output from quarto_plot_exception
-    error_patterns = [
-        ("Error rendering", "Plot rendering error"),
-        ("Traceback (most recent call last)", "Python traceback"),
-        ("ValueError:", "ValueError exception"),
-        ("TypeError:", "TypeError exception"),
-        ("KeyError:", "KeyError exception"),
-        ("AttributeError:", "AttributeError exception"),
-        ("NameError:", "NameError exception"),
-        ("Exception:", "Generic exception"),
-        ("The given query resulted in an empty dataframe", "Empty dataframe error"),
-    ]
-
-    for pattern, description in error_patterns:
-        if pattern in content:
-            count = content.count(pattern)
-            if count > 1:
-                errors.append(f"{description} (found {count} times)")
-            else:
-                errors.append(description)
-
-    return errors
