+++ python/pdstools/utils/report_utils/_common.py
@@ -1,0 +1,5 @@
+"""Shared module-level state for the ``report_utils`` submodules."""
+
+import logging
+
+logger = logging.getLogger("pdstools.utils.report_utils")
